refactor: moving datepicker type to new pattern
