+++ stories/Datepicker.stories.ts
@@ -1,9 +1,6 @@
 import { Meta, Story } from '@storybook/angular/types-6-0';
 import { action } from '@storybook/addon-actions';
-import {
-  IonDatepickerComponent,
-  IonDatePickerComponentProps,
-} from './../projects/ion/src/lib/picker/date-picker/date-picker.component';
+import { IonDatepickerComponent } from './../projects/ion/src/lib/picker/date-picker/date-picker.component';
 import { IonDatePickerCalendarComponent } from './../projects/ion/src/lib/picker/date-picker/date-picker-calendar/date-picker-calendar.component';
 import { IonDatePickerInputComponent } from './../projects/ion/src/lib/picker/date-picker/date-picker-input/date-picker-input.component';
 import { IonControlPickerComponent } from './../projects/ion/src/lib/picker/control-picker/control-picker.component';
@@ -13,6 +10,7 @@ import { IonDividerModule } from '../projects/ion/src/lib/divider/divider.module
 import { IonInputModule } from '../projects/ion/src/lib/input/input.module';
 import { IonTooltipComponent } from '../projects/ion/src/lib/tooltip/tooltip.component';
 import { IonTooltipModule } from '../projects/ion/src/public-api';
+import { IonDatePickerComponentProps } from 'ion/lib/core/types/datepicker';
 
 export default {
   title: 'Ion/Data Entry/Datepicker',
